requirements.txt
aiohttp>=3.8.4
aiosignal>=1.3.1
async-timeout>=4.0.2
attrs>=23.1.0
black>=23.7.0
certifi>=2023.5.7
cffi>=1.15.1
charset-normalizer>=3.2.0
click>=8.1.5
colorama>=0.4.6
decorator>=5.1.1
discord>=2.3.1
discord.py>=2.3.1
frozenlist>=1.4.0
idna>=3.4
imageio>=2.31.1
imageio-ffmpeg>=0.4.8
moviepy>=1.0.3
multidict>=6.0.4
mutagen>=1.46.0
mypy-extensions>=1.0.0
numpy>=1.25.1
packaging>=23.1
pathspec>=0.11.1
Pillow>=10.0.0
platformdirs>=3.9.1
proglog>=0.1.10
pycparser>=2.21
PyNaCl>=1.5.0
python-dotenv>=1.0.0
pytube>=15.0.0
requests>=2.31.0
tqdm>=4.65.0
urllib3>=2.0.3
yarl>=1.9.2
toml>=0.10.2
gspread